+++ src/main/java/peaksoft/giftlistj7g2/controller/AuthController.java
@@ -3,14 +3,11 @@ package peaksoft.giftlistj7g2.controller;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import peaksoft.giftlistj7g2.model.dto.LoginResponse;
-import peaksoft.giftlistj7g2.model.dto.RegisterRequest;
-import peaksoft.giftlistj7g2.model.dto.RegisterResponse;
 import peaksoft.giftlistj7g2.model.dto.LoginRequest;
-import peaksoft.giftlistj7g2.model.service.UserService;
+import peaksoft.giftlistj7g2.model.service.AuthService;
 
 
 @RestController
@@ -18,18 +15,10 @@ import peaksoft.giftlistj7g2.model.service.UserService;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class AuthController {
     @Autowired
-    UserService userService;
-
-    @PostMapping("/sign-up")
-    public ResponseEntity<RegisterResponse> registration(@RequestBody RegisterRequest request) {
-        RegisterResponse response = userService.save(request);
-        return new ResponseEntity<>(response, HttpStatus.CREATED);
-    }
-
+    AuthService userService;
 
     @PostMapping("/sign-in")
     public ResponseEntity<LoginResponse> singIn(@RequestBody LoginRequest request) {
-        System.out.println("Log in controller");
         LoginResponse response = userService.login(request);
         return ResponseEntity.ok(response);
     }
