+++ src/main/java/peaksoft/giftlistj7g2/model/dto/LoginRequest.java
@@ -1,11 +1,14 @@
 package peaksoft.giftlistj7g2.model.dto;
 
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
+import lombok.experimental.FieldDefaults;
 
 @Getter
 @Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
 public class LoginRequest {
-    private String email;
-    private String password;
+    String email;
+    String password;
 }
