--- src/main/java/peaksoft/giftlistj7g2/model/dto/RegisterRequest.java
@@ -1,18 +1,0 @@
-package peaksoft.giftlistj7g2.model.dto;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class RegisterRequest {
-    String name;
-    String lastName;
-    String email;
-    int age;
-    String password;
-}
-
