--- src/main/java/peaksoft/giftlistj7g2/model/dto/RegisterResponse.java
@@ -1,20 +1,0 @@
-package peaksoft.giftlistj7g2.model.dto;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@Builder
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class RegisterResponse {
-    Long id;
-    String name;
-    String lastName;
-    String email;
-    int age;
-    String response;
-}
