--- src/main/java/peaksoft/giftlistj7g2/model/dto/UserRequest.java
@@ -1,16 +1,0 @@
-package peaksoft.giftlistj7g2.model.dto;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class UserRequest {
-    String name;
-    String lastname;
-    String email;
-    int age;
-}
