--- src/main/java/peaksoft/giftlistj7g2/model/dto/UserResponse.java
@@ -1,22 +1,0 @@
-package peaksoft.giftlistj7g2.model.dto;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-import java.time.LocalDate;
-
-@Getter
-@Setter
-@Builder
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class UserResponse {
-    Long id;
-    String name;
-    String lastname;
-    String email;
-    int age;
-    LocalDate createDate;
-}
