+++ src/main/java/peaksoft/giftlistj7g2/model/repository/AuthRepository.java
@@ -9,7 +9,7 @@ import peaksoft.giftlistj7g2.model.entities.User;
 import java.util.Optional;
 
 @Repository
-public interface AuthRepository extends JpaRepository<User, Long>{
+public interface AuthRepository extends JpaRepository<User, Long> {
     @Query("select user from User user where user.email=:email")
     Optional<User> findByEmail(@Param("email") String email);
 }
